Fix Python 3.7/3.8 tests
